add dashboard examples in new tab style
