+++ packages/ui/src/lib/layoutExamples/Dashboards/demoTabs/index.ts
@@ -1,0 +1,4 @@
+export { default as Demo1 } from './Demo1.svelte';
+export { default as Leggi } from './Leggi.svelte';
+export { default as Pop1 } from './Pop1.svelte';
+export { default as Sewers } from './Sewers.svelte';
