+++ packages/ui/src/lib/sidebar/Sidebar.svelte
@@ -137,7 +137,7 @@
 						{#if tabs.length}
 							{#each tabs as tab}
 								{#if component && selectedValue === tab.id}
-									<TabPanel tabPanelId={`${tab.id}-panel`} tabId={tab.id}>
+									<TabPanel tabPanelId={`${tab.id}-panel`} tabId={tab.id} class="space-y-4">
 										<svelte:component this={component} />
 									</TabPanel>
 								{/if}
